requirements.txt
astroid==2.2.5
backcall==0.1.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
gunicorn==19.9.0
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.16
jedi==0.13.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1
yapf==0.26.0
